RobotMovement with Light
I've added a new variable called Robots which is a list of both robots. The robots coordinates will be sent to the LightCommand function. The LightCommand function will choose a random colour for the lights and w1 and z1 and w2 and y2 will be the previous movements and.the red light should make the coordinates unchangable.